requirements.txt
mattermostdriver
psycopg
python-dotenv 